requirements.txt
playwright==1.49.1
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
opencv-python==4.10.0.84
Pillow==11.1.0
python-telegram-bot==22.5
python-dotenv==1.1.0
google-genai==1.47.0
